Generalize node input data as either constants, batch buffers, custom sized buffers or interpolated parameters. Make some integrations. Improved node graph test so it smoke tests all existing node operations.

+++ packages/nodegraph/source/common/core/NodeGraphOutput.cpp
@@ -17,7 +17,7 @@ namespace l::nodegraph {
         }
         if (static_cast<int32_t>(mOutputBuf->size()) < size) {
             mOutputBuf->resize(size);
-            for (int32_t i = 0; i < mOutputBuf->size(); i++) {
+            for (size_t i = 0; i < mOutputBuf->size(); i++) {
                 (*mOutputBuf)[i] = 0.0f;
             }
         }
